Conflicts:
	doc/er/LiveCart.dez

+++ application/model/order/CustomerOrder.php
@@ -49,6 +49,8 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 
 	private $isRulesProcessed;
 
+	private static $isEmptyAllowed = false;
+
 	const STATUS_NEW = 0;
 	const STATUS_PROCESSING = 1;
 	const STATUS_AWAITING = 2;
@@ -111,6 +113,11 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 	public static function getInstanceById($id, $loadData = self::LOAD_DATA, $loadReferencedRecords = false)
 	{
 		return parent::getInstanceById('CustomerOrder', $id, $loadData, $loadReferencedRecords);
+	}
+
+	public static function getInstanceByInvoiceNumber($id, $loadReferencedRecords = false)
+	{
+		return self::getRecordSet(select(eq(f('CustomerOrder.invoiceNumber'), $id)), $loadReferencedRecords)->shift();
 	}
 
 	/**
@@ -181,7 +188,7 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 
 		if (!$this->isFinalized->get() && $this->orderedItems)
 		{
-			$this->updateToStock();
+			return $this->updateToStock();
 		}
 
 		$this->event('after-load');
@@ -189,14 +196,18 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 
 	public function loadAddresses()
 	{
-		if ($this->billingAddress->get())
+		$billingAddress = $this->billingAddress->get();
+		if ($billingAddress)
 		{
-			$this->billingAddress->get()->load(self::LOAD_REFERENCES);
+			$billingAddress->load(self::LOAD_REFERENCES);
+			$billingAddress->getSpecification(); // todo: why EavObject not loaded automaticaly?
 		}
 
-		if ($this->shippingAddress->get())
+		$shippingAddress = $this->shippingAddress->get();
+		if ($shippingAddress)
 		{
-			$this->shippingAddress->get()->load(self::LOAD_REFERENCES);
+			$shippingAddress->load(self::LOAD_REFERENCES);
+			$shippingAddress->getSpecification();
 		}
 	}
 
@@ -452,8 +463,13 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 	 *
 	 *  @return CustomerOrder New order instance containing wishlist items
 	 */
-	public function finalize()
+	public function finalize($options = array())
 	{
+		if ($this->isFinalized->get() && empty($options['allowRefinalize']))
+		{
+			return;
+		}
+
 		self::beginTransaction();
 
 		$this->event('before-finalize');
@@ -463,6 +479,11 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 
 		foreach ($this->getShipments() as $shipment)
 		{
+			if ($shipment->isExistingRecord())
+			{
+				$shipment->deleteRecordSet('ShipmentTax', new ARDeleteFilter());
+			}
+
 			$shipment->order->set($this);
 			$shipment->save();
 
@@ -479,7 +500,11 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 
 		foreach ($this->getShoppingCartItems() as $item)
 		{
-			$item->price->set($item->getItemPrice());
+			if (!empty($options['customPrice']))
+			{
+				$item->price->set($item->getSubTotalBeforeTax() / $item->getCount());
+			}
+
 			$item->name->set($item->getProduct()->getParent()->name->get());
 			$item->setValueByLang('name', 'sku', $item->getProduct()->sku->get());
 			$item->save();
@@ -503,35 +528,47 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 			}
 		}
 
-		if (!$this->shippingAddress->get() && $this->user->get()->defaultShippingAddress->get())
+		if (!$this->shippingAddress->get() && $this->user->get() && $this->user->get()->defaultShippingAddress->get() && $this->isShippingRequired())
 		{
 			$this->shippingAddress->set($this->user->get()->defaultShippingAddress->get()->userAddress->get());
 		}
 
-		if (!$this->billingAddress->get() && $this->user->get()->defaultBillingAddress->get())
+		if (!$this->billingAddress->get() && $this->user->get() && $this->user->get()->defaultBillingAddress->get())
 		{
 			$this->billingAddress->set($this->user->get()->defaultBillingAddress->get()->userAddress->get());
 		}
 
 		// clone billing/shipping addresses
-		foreach (array('billingAddress', 'shippingAddress') as $address)
+		if (!$this->isFinalized->get())
 		{
-			if ($this->$address->get())
+			foreach (array('billingAddress', 'shippingAddress') as $address)
 			{
-				$this->$address->get()->getSpecification();
-				$cloned = clone $this->$address->get();
-				$cloned->save();
-				$this->$address->set($cloned);
+				if ($this->$address->get())
+				{
+					$this->$address->get()->load();
+					$this->$address->get()->getSpecification();
+					$cloned = clone $this->$address->get();
+					$cloned->save();
+					$cloned->loadEav();
+					$this->$address->set($cloned);
+				}
 			}
 		}
 
 		// move wish list items to a separate order
-		$wishList = CustomerOrder::getNewInstance($this->user->get());
-		foreach ($this->getWishListItems() as $item)
+		if ($this->getWishListItems())
 		{
-			$wishList->addItem($item);
+			$wishList = CustomerOrder::getNewInstance($this->user->get());
+			foreach ($this->getWishListItems() as $item)
+			{
+				$wishList->addItem($item);
+			}
+			$wishList->save();
 		}
-		$wishList->save();
+		else
+		{
+			$wishList = null;
+		}
 
 		// set order total
 		$this->totalAmount->set($this->getTotal(true));
@@ -561,18 +598,21 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 		$shipments = $this->shipments;
 		unset($this->shipments);
 
-		$generator = InvoiceNumberGenerator::getGenerator($this);
-		$saved = false;
-		while (!$saved)
+		if (!$this->invoiceNumber->get())
 		{
-			try
+			$generator = InvoiceNumberGenerator::getGenerator($this);
+			$saved = false;
+			while (!$saved)
 			{
-				$this->invoiceNumber->set($generator->getNumber());
-				$this->save();
-				$saved = true;
-			}
-			catch (SQLException $e)
-			{
+				try
+				{
+					$this->invoiceNumber->set($generator->getNumber());
+					$this->save();
+					$saved = true;
+				}
+				catch (SQLException $e)
+				{
+				}
 			}
 		}
 
@@ -715,6 +755,11 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 
 	public function setUser(User $user)
 	{
+		if ($this->user->get() && ($this->user->get()->getID() == $user->getID()))
+		{
+			return;
+		}
+
 		$this->user->set($user);
 		$this->setCheckoutStep(self::CHECKOUT_USER);
 
@@ -738,13 +783,18 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 
 	public function setCheckoutStep($step)
 	{
-		if ($step < $this->checkoutStep->get())
+		if ($step <= $this->checkoutStep->get())
 		{
 			return false;
 		}
 
 		$this->checkoutStep->set($step);
 		$this->save();
+	}
+
+	public static function allowEmpty($allow = true)
+	{
+		self::$isEmptyAllowed = $allow;
 	}
 
 	/*####################  Saving ####################*/
@@ -836,13 +886,20 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 
 			$this->totalAmount->set($this->getTotal(true));
 		}
+		else
+		{
+			if (!$this->isShippingRequired())
+			{
+				//$this->shippingAddress->setNull();
+			}
+		}
 
 		if ($this->isModified() || $isModified)
 		{
 			$this->serializeShipments();
 		}
 
-		if (!$this->isFinalized->get() && !$this->orderedItems && !$allowEmpty)
+		if (!$this->isFinalized->get() && !$this->orderedItems && !$allowEmpty && !self::$isEmptyAllowed)
 		{
 			$this->delete();
 			return false;
@@ -1322,13 +1379,10 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 		foreach ($this->getOrderedItems() as $item)
 		{
 			$product = $item->getProduct();
-			if (!$product)
+			if (!$product || (!$product->isEnabled->get() || !$product->getParent()->isEnabled->get()))
 			{
 				$this->removeItem($item);
-			}
-			else if (!$product->isEnabled->get() || !$product->getParent()->isEnabled->get())
-			{
-				$this->removeItem($item);
+				$result['delete'][] = $item->toArray();
 			}
 		}
 
@@ -1350,16 +1404,19 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 
 			if (!$product->isBackOrderable->get() && !$item->isSavedForLater->get() && !$product->isBundle())
 			{
-				if ($product->stockCount->get() <= 0)
+				if (!$product->isDownloadable() || $product->isInventoryTracked())
 				{
-					$item->isSavedForLater->set(OrderedItem::OUT_OF_STOCK);
-					$result['out'][] = array('id' => $item->getID());
-				}
-				else if ($product->stockCount->get() < $item->count->get())
-				{
-					$count = $item->count->get();
-					$item->count->set($product->stockCount->get());
-					$result['count'][] = array('id' => $item->getID(), 'from' => $count, 'to' => $item->count->get());
+					if (($product->stockCount->get() <= 0))
+					{
+						$item->isSavedForLater->set(OrderedItem::OUT_OF_STOCK);
+						$result['out'][] = array('id' => $item->getID());
+					}
+					else if ($product->stockCount->get() < $item->count->get())
+					{
+						$count = $item->count->get();
+						$item->count->set($product->stockCount->get());
+						$result['count'][] = array('id' => $item->getID(), 'from' => $count, 'to' => $item->count->get());
+					}
 				}
 			}
 		}
@@ -1619,6 +1676,7 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 
 		// shipping subtotal
 		$array['shippingSubtotal'] = null;
+		$array['shippingSubtotalWithoutTax'] = null;
 		if ($this->shipments)
 		{
 			foreach ($this->shipments as $shipment)
@@ -1627,11 +1685,12 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 				if (!is_null($shipmentShipping))
 				{
 					$array['shippingSubtotal'] += $shipment->getShippingTotalWithTax();
+					$array['shippingSubtotalWithoutTax'] += $shipment->getShippingTotalBeforeTax();
 				}
 			}
 		}
 
-		$array['subtotalBeforeTaxes'] = $array['itemSubtotalWithoutTax'] + $array['shippingSubtotal'];
+		$array['subtotalBeforeTaxes'] = $array['itemSubtotalWithoutTax'] + $array['shippingSubtotalWithoutTax'];
 
 		foreach (array('amountPaid', 'amountNotCaptured', 'amountDue', 'itemSubtotal', 'shippingSubtotal', 'subtotalBeforeTaxes', 'totalAmount', 'itemDiscountReverse', 'itemDiscount', 'itemSubtotalWithoutTax') as $key)
 		{
@@ -1643,6 +1702,7 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 
 		if (!$array['isFinalized'])
 		{
+			$this->isRulesProcessed = false;
 			$isOrderable = $this->isOrderable();
 			if ($isOrderable instanceof OrderException)
 			{
@@ -1652,8 +1712,13 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 			$array['isOrderable'] = !($isOrderable instanceof OrderException) && $isOrderable;
 
 			$array['isShippingSelected'] = $this->isShippingSelected();
-			$array['isShippingSelected'] = $this->isShippingSelected();
 			$array['isAddressSelected'] = ($this->shippingAddress->get() && $this->billingAddress->get());
+		}
+
+		// otherwise left empty on payment page for some reason...
+		if ($this->billingAddress->get())
+		{
+			$array['BillingAddress'] = $this->billingAddress->get()->toArray();
 		}
 
 		$this->setArrayData($array);
@@ -1924,6 +1989,16 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 
 	public function processBusinessRules($reload = false)
 	{
+		if ($this->isFinalized->get())
+		{
+			return;
+		}
+
+		foreach ($this->getShoppingCartItems() as $item)
+		{
+			$item->reset();
+		}
+
 		foreach ($this->getDiscountActions($reload) as $ruleAction)
 		{
 			if ($ruleAction->isOrderAction())
@@ -1980,6 +2055,11 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 		}
 
 		return $this->deliveryZone;
+	}
+
+	public function setDeliveryZone(DeliveryZone $zone)
+	{
+		$this->deliveryZone = $zone;
 	}
 
 	/**
@@ -2090,7 +2170,7 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 		ProductPrice::loadPricesForRecordSet($set);
 	}
 
-	protected function __get($name)
+	public function __get($name)
 	{
 		switch ($name)
 		{
@@ -2119,23 +2199,33 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 		$this->shipments = new ARSet();
 		$this->orderedItems = array();
 
-		foreach ($original->getShipments() as $shipment)
-		{
-			$cloned = clone $shipment;
-			$cloned->order->set($this);
+		if ($original->isFinalized->get())
+        {
+                foreach ($original->getShipments() as $shipment)
+                {
+                        $cloned = clone $shipment;
+                        $cloned->order->set($this);
 
-			if ($this->isMultiAddress->get())
-			{
-				$this->addShipment($cloned);
-			}
-			else
-			{
-				foreach ($cloned->getItems() as $item)
-				{
-					$this->addItem($item);
-				}
-			}
-		}
+                        if ($this->isMultiAddress->get())
+                        {
+                                $this->addShipment($cloned);
+                        }
+                        else
+                        {
+                                foreach ($cloned->getItems() as $item)
+                                {
+                                        $this->addItem($item);
+                                }
+                        }
+                }
+        }
+        else
+        {
+                foreach ($original->getOrderedItems() as $item)
+                {
+                        $this->addItem(clone $item);
+                }
+        }
 
 		if ($this->isMultiAddress->get())
 		{
@@ -2168,7 +2258,6 @@ class CustomerOrder extends ActiveRecordModel implements EavAble, BusinessRuleOr
 		{
 			$this->shippingAddress->set($this->getClonedAddress($this->shippingAddress->get(), false));
 		}
-
 		$this->save();
 	}
